Moved terraform versions filter to lib; added test to make sure versions are always in the correct order
